+++ CHANGELOG.md
@@ -12,6 +12,8 @@ in this document.
 15.2.x
 ------
 
+* Added `aiomisc.service.ProcessService` for running python function in
+  separate process.
 * Change order of detection for `aiomisc_log.LogFormat.default()` in case
   the stderr is not a tty. Fixes `journald` formatter auto detection.
 * Rewritten `aiomisc.iterator_wrapper.IteratorWrapper`  to avoid unnecessary